user-workspace/snapstory/ai_service/requirements.txt
flask==2.3.2
flask-cors==3.0.10
tensorflow==2.12.0
transformers==4.30.2
beautifulsoup4==4.12.2
selenium==4.10.0
pymongo==4.3.3
python-dotenv==1.0.0